I optimized matmul.
(Optimization may be still possible.)

before
```
TotalRunTime,303490,  sum:303.49ms
kn2row,166129,  sum:166.129ms
kn2row-1x1,2094,  sum:2.094ms
kn2row-buf,4,  sum:0.004ms
matrix_multiplication,108444,2085,  sum:110.529ms
```
after
```
TotalRunTime,297751,  sum:297.751ms
kn2row,160168,  sum:160.168ms
kn2row-1x1,2087,  sum:2.087ms
kn2row-buf,4,  sum:0.004ms
matrix_multiplication,102486,2078,  sum:104.564ms
```