@lm-konda 2018-11-23
I did simple optimization.(GCC optimization is very weak...)

before
```
TotalRunTime,297807,  sum:297.807ms
kn2row,160437,  sum:160.437ms
kn2row-1x1,2090,  sum:2.09ms
kn2row-buf,4,  sum:0.004ms
matrix_multiplication,102152,2080,  sum:104.232ms
```
after
```
TotalRunTime,294501,  sum:294.501ms
kn2row,157131,  sum:157.131ms
kn2row-1x1,2091,  sum:2.091ms
kn2row-buf,5,  sum:0.005ms
matrix_multiplication,98972,2082,  sum:101.054ms
```